backend/requirements.txt
tensorflow_hub
pandas
scikit-image
numpy
opencv-python
tensorflow
flask
cv2
flask_cors
google.generativeai
python-dotenv
pymongo[srv]
bcrypt